proxy/requirements.txt
scrapy
scrapy_redis
zerorpc
requests
python-crontab
psutil